requirements.txt
# MedConnect - Dependencias de Python
# Para despliegue en Railway con Flask, Google Sheets y Telegram Bot

# Core dependencies
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Google Sheets integration
gspread==5.12.4
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP requests and API
requests==2.31.0
urllib3==2.1.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Security and authentication
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.8

# File handling
Werkzeug==3.0.1
Pillow==10.1.0

# Date and time handling
python-dateutil==2.8.2

# Utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Railway deployment
whitenoise==6.6.0

# Flask y extensiones
Flask-CORS==4.0.0

# Natural Language Processing básico
spacy==3.7.2
nltk==3.8.1

# Web Framework para APIs REST
flask-cors==4.0.0

# Fecha y hora
python-logging==0.1.0

# PDF Generation
reportlab==4.0.7 